+++ Core/Core/API/APIConversation.swift
@@ -65,114 +65,6 @@ struct APIConversationMessage: Codable, Equatable {
     let forwarded_messages: [APIConversationMessage]?
 }
 
-struct GetConversationsRequest: APIRequestable {
-    typealias Response = [APIConversation]
-    enum Include: String {
-        case participant_avatars
-    }
-    enum Scope: String {
-        case unread, starred, archived, sent
-    }
-
-    let path = "conversations"
-
-    let perPage: Int?
-    let include: [Include]
-    let scope: Scope?
-
-    var query: [APIQueryItem] {
-        var query: [APIQueryItem] = [
-            .include(include.map { $0.rawValue }),
-        ]
-        if let perPage = perPage {
-            query.append(.value("per_page", String(perPage)))
-        }
-        if let scope = scope {
-            query.append(.value("scope", scope.rawValue))
-        }
-        return query
-    }
-
-    init(include: [Include] = [], perPage: Int? = nil, scope: Scope? = nil) {
-        self.include = include
-        self.perPage = perPage
-        self.scope = scope
-    }
-}
-
-struct GetConversationRequest: APIRequestable {
-    typealias Response = APIConversation
-    enum Include: String {
-        case participant_avatars
-    }
-
-    let id: String
-    let include: [Include]
-    var path: String { "conversations/\(id)" }
-    var query: [APIQueryItem] {
-        return [
-            .include(include.map { $0.rawValue }),
-        ]
-    }
-
-    init(id: String, include: [Include] = []) {
-        self.id = id
-        self.include = include
-    }
-}
-
-struct PutConversationRequest: APIRequestable {
-    typealias Response = APIConversation
-    struct Body: Encodable {
-        let id: String
-        let workflow_state: ConversationWorkflowState
-    }
-
-    let id: String
-    let workflowState: ConversationWorkflowState
-    var path: String { "conversations/\(id)" }
-    let method = APIMethod.put
-    var body: Body? {
-        return Body(id: id, workflow_state: workflowState)
-    }
-}
-
-struct PostAddMessageRequest: APIRequestable {
-    typealias Response = APIConversation
-    struct Body: Encodable {
-        let recipients: [String]
-        let body: String
-        let subject: String?
-        let attachment_ids: [String]?
-        let media_comment_id: String?
-        let media_comment_type: MediaCommentType?
-        let context_code: String?
-        let sendIndividually: Bool
-
-        enum CodingKeys: CodingKey {
-            case recipients, body, subject, group_conversation, attachment_ids, media_comment_id, media_comment_type, context_code, bulk_message
-        }
-
-        func encode(to encoder: Encoder) throws {
-            var container = encoder.container(keyedBy: CodingKeys.self)
-            try container.encode(recipients, forKey: .recipients)
-            try container.encode(body, forKey: .body)
-            try container.encode(subject, forKey: .subject)
-            try container.encode(true, forKey: .group_conversation)
-            try container.encode(attachment_ids, forKey: .attachment_ids)
-            try container.encode(media_comment_id, forKey: .media_comment_id)
-            try container.encode(media_comment_type, forKey: .media_comment_type)
-            try container.encode(context_code, forKey: .context_code)
-            if sendIndividually {
-                try container.encode(1, forKey: .bulk_message)
-            }
-        }
-    }
-    let id: String
-    let message: Body
-    var path: String { "conversations/\(id)/add_message" }
-}
-
 #if DEBUG
 extension APIConversation {
     static func make(
